requirements.txt
PySide6
PySide6-Fluent-Widgets
PyYAML
numpy
pandas
openpyxl
